requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
crispy-bootstrap5==0.7
Django==4.2.6
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.1.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
